moved component folders into component directory
