+++ instagram-app/src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import logo from './logo.svg';
 import instagram from './instagram.svg';
 import './App.css';
-import SearchBar from './SearchBar/SearchBar';
+import SearchBar from './components/SearchBar/SearchBar';
 import dummyData from './dummy-data';
 import PostContainer from './PostContainer/PostContainer';
